requirements.txt
requests==2.32.5
python-dotenv==1.2.1
supabase==2.4.0
# httpx 0.27.2 is required for gotrue proxy param, but supabase 2.4.0
# declares httpx<0.26. It works fine — just causes a pip warning.
# Upgrade supabase to 2.27+ when pyroaring/MSVC build issue is resolved.
httpx==0.27.2
beautifulsoup4==4.12.3
pytrends==4.9.2
anthropic==0.43.0
